Fix some Sound extensions
Sound vomume scale in 0-1.0f was mixed with 0-255 scale

+++ Quake/q_sound.h
@@ -93,7 +93,7 @@ void S_Init (void);
 void S_Startup (void);
 void S_Shutdown (void);
 void S_StartSound (int entnum, int entchannel, sfx_t *sfx, vec3_t origin, float fvol, float attenuation);
-void S_StaticSound (sfx_t *sfx, vec3_t origin, float vol, float attenuation);
+void S_StaticSound (sfx_t *sfx, vec3_t origin, int vol, float attenuation);
 void S_StopSound (int entnum, int entchannel);
 void S_StopAllSounds (qboolean clear, qboolean keep_statics);
 void S_ClearBuffer (void);
